Updated discunt api for more accurate representation of price.

+++ Backend/Car_Rental_System/Controllers/CarsController.cs
@@ -135,7 +135,7 @@ namespace Car_Rental_System.Controllers
 
         //add discount to car
         [HttpPut("addDiscount")]
-        public async Task<IActionResult> AddDiscount(decimal discount, string carId)
+        public async Task<IActionResult> AddDiscount(double discount, string carId)
         {
             var car = await dbContext.Cars.FindAsync(Guid.Parse(carId));
             if (car == null)
